+++ src/senaite/storage/browser/container/view.py
@@ -51,7 +51,7 @@ class ContainerListingView(FacilityListingView):
 
         self.context_actions = collections.OrderedDict((
             (_("Add container"), {
-                "url": "createObject?type_name=StorageContainer",
+                "url": "++add++StorageContainer",
                 "permission": AddStorageContainer,
                 "icon": "{}/{}".format(
                     self.icon_path, "storage-container"),